fix get information user login to order

+++ database/migrations/2025_05_17_021828_create_coupons_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->string('code')->unique();                     // Mã giảm giá
             $table->unsignedInteger('discount');                      // Giá trị giảm cố định (VD: 50.000đ)
-            $table->decimal('min_order_amount', 10, 2)->nullable(); // Giá trị đơn hàng tối thiểu
+            $table->unsignedInteger('min_order_amount')->nullable(); // Giá trị đơn hàng tối thiểu
 
             $table->unsignedInteger('usage_limit')->nullable();   // Tổng số lượt dùng toàn hệ thống
             $table->unsignedInteger('user_limit')->nullable();    // Số lượt dùng mỗi user
